refactor(web-ui): rename http client module
