+++ packages/frontend/web-ui/src/app/store/thunk/createAuthByToken.spec.ts
@@ -1,12 +1,12 @@
 import { afterAll, beforeAll, describe, expect, it, jest } from '@jest/globals';
 
-jest.mock('../../../common/http/services/HttpService');
+jest.mock('../../../common/http/services/httpClient');
 jest.mock('../../../common/http/helpers/buildSerializableResponse');
 
 import { buildSerializableResponse } from '../../../common/http/helpers/buildSerializableResponse';
 import { AuthResponse } from '../../../common/http/models/AuthResponse';
 import { AuthSerializedResponse } from '../../../common/http/models/AuthSerializedResponse';
-import { httpClient } from '../../../common/http/services/HttpService';
+import { httpClient } from '../../../common/http/services/httpClient';
 import { createAuthByToken } from './createAuthByToken';
 
 describe(createAuthByToken.name, () => {
