+++ packages/frontend/web-ui/src/app/store/thunk/createAuthByToken.ts
@@ -3,7 +3,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { buildSerializableResponse } from '../../../common/http/helpers/buildSerializableResponse';
 import { AuthResponse } from '../../../common/http/models/AuthResponse';
 import { AuthSerializedResponse } from '../../../common/http/models/AuthSerializedResponse';
-import { httpClient } from '../../../common/http/services/HttpService';
+import { httpClient } from '../../../common/http/services/httpClient';
 
 export const createAuthByToken = createAsyncThunk(
   'auth/createAuthByToken',
