+++ packages/frontend/web-ui/src/app/store/thunk/getUserMe.spec.ts
@@ -1,12 +1,12 @@
 import { afterAll, beforeAll, describe, expect, it, jest } from '@jest/globals';
 
-jest.mock('../../../common/http/services/HttpService');
+jest.mock('../../../common/http/services/httpClient');
 jest.mock('../../../common/http/helpers/buildSerializableResponse');
 
 import { buildSerializableResponse } from '../../../common/http/helpers/buildSerializableResponse';
 import { UserMeResponse } from '../../../common/http/models/UserMeResponse';
 import { UserMeSerializedResponse } from '../../../common/http/models/UserMeSerializedResponse';
-import { httpClient } from '../../../common/http/services/HttpService';
+import { httpClient } from '../../../common/http/services/httpClient';
 import { getUserMe } from './getUserMe';
 
 describe(getUserMe.name, () => {
