+++ packages/frontend/web-ui/src/app/store/thunk/getUserMe.ts
@@ -3,7 +3,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { buildSerializableResponse } from '../../../common/http/helpers/buildSerializableResponse';
 import { UserMeResponse } from '../../../common/http/models/UserMeResponse';
 import { UserMeSerializedResponse } from '../../../common/http/models/UserMeSerializedResponse';
-import { httpClient } from '../../../common/http/services/HttpService';
+import { httpClient } from '../../../common/http/services/httpClient';
 
 export const getUserMe = createAsyncThunk(
   'user/getUserMe',
