+++ packages/frontend/web-ui/src/common/http/services/__mocks__/httpClient.ts
